requirements.txt
Pillow==9.5.0
requests==2.26.0
